+++ src/app/dashboard/account-edit.component.ts
@@ -14,6 +14,7 @@ export class AccountEditComponent implements OnInit {
   accountType: any;
   // @ViewChild('fileInput') fileInput;
   submitStatus: any;
+  submitStatusError: string = 'Error received, please try again.';
 
   constructor(
   private http: Http,
@@ -111,7 +112,16 @@ export class AccountEditComponent implements OnInit {
         console.log(this.submitStatus);
       },
       error => {
-        console.log( error._body );
+        console.log('Edit Error');
+        console.log(error);
+        try {
+          console.log(error.error);
+          let jsonError = error.json();
+          console.log("boop");
+          this.submitStatusError = '"' + jsonError.error + '" Error, ' + jsonError.message;
+        } catch(e) {
+          this.submitStatusError = 'There was a server error, please try again later.';
+        }
         this.submitStatus = "send_failed";
         console.log(this.submitStatus);
       }
@@ -159,7 +169,16 @@ export class AccountEditComponent implements OnInit {
         console.log(this.submitStatus);
       },
       error => {
-        console.log( error._body );
+        console.log('Edit Error');
+        console.log(error);
+        try {
+          console.log(error.error);
+          let jsonError = error.json();
+          console.log("boop");
+          this.submitStatusError = '"' + jsonError.error + '" Error, ' + jsonError.message;
+        } catch(e) {
+          this.submitStatusError = 'There was a server error, please try again later.';
+        }
         this.submitStatus = "send_failed";
         console.log(this.submitStatus);
       }