+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 # Next Release
 
+* Fixed register box view on login page
+* Made the form reset properly if adding a new one
+* Added better error messages on register and account edit
+
 # v0.0.1
 
 * Added Travis Testing